Fixed a bug where you couldn't upload an avatar if you didn't already have one.

+++ server/api/v1/user/[id]/import-avatar.put.mjs
@@ -7,7 +7,9 @@ export default defineEventHandler(async (event) => {
     const users = await useKind('User');
     const { id } = getRouterParams(event);
     const user = await users.findOrDie(id);
-    const oldAvatar = await user.parent('avatar');
+    let oldAvatar = undefined;
+    if (user.avatarId)
+        oldAvatar = await user.parent('avatar');
     const session = obtainSession(event);
     await user.canEdit(session);
     const body = await getBody(event);
@@ -18,6 +20,7 @@ export default defineEventHandler(async (event) => {
     await s3file.verifyExactDimensions(300, 300);
     user.set('avatarId', s3file.get('id'));
     await user.update();
-    await oldAvatar.delete();
+    if (oldAvatar)
+        await oldAvatar.delete();
     return await user.describe(describeParams(event, session));
 });
